Modified the schema validation method (Check Format) to first use the name space url defined in the xml and then fall back to the imdi schema and then the schema provided in the jar file. Added clarin XSD reader. Loaded Clarin templates from the schema location of cmdi files. Addded component insertion into clarin files (via add menu). Added auto recovery from history when a file is missing. Updated the add menu contents to handle clarin sub nodes.

+++ src/nl/mpi/arbil/data/ImdiTreeObject.java
@@ -413,7 +413,7 @@ public class ImdiTreeObject implements Comparable {
                     Document nodDom = null;
                     // cacheLocation will be null if useCache = false hence no file has been saved
 //        String cacheLocation = null;
-                    if (this.isLocal() && !this.getFile().exists()) {
+                    if (this.isLocal() && !this.getFile().exists() && new File(this.getFile().getAbsolutePath() + ".0").exists()) {
                         // if the file is missing then try to find a valid history file
                         try {
                             FileOutputStream outFile = new FileOutputStream(this.getFile());
